Modify PlayerMain
Add Jump, IsGrounded, CreateRays, VelocityDecay,
and OnCollisionEnter2D methods.

Jump applies a upward force to player rigid body.
This only occurs when IsGrounded returns true and
_isAirborne is false. After applying the upward
force, _isAirborne is set to true to prevent
the player from jumping in midair.

IsGrounded checks whether the player is over solid ground.
It calls CreateRays to create 3 rays at the base of the
player sprite, at the leftmost, right most and center
position horizontal positions of the sprite. If any of
the raycasts detect a collision with a gameobject in the
Ground layer, IsGrounded returns true. Else, returns false.

VelocityDecay reduces player's horizontal and vertical
velocity over time.

OnCollisionEnter2D currently checks for an upward
force collision i.e. the player lands on the ground. If
this is true, then _isAirborne is set to false. This
will allow the player to jump again.