Add LevelManager and DimensionSwitch classes Modify PlayerMain
LevelManager is responsible for getting the
current level's elements. Currently, it only
has the SwitchDimensions method that prints
out a debug message.

DimensionSwitch will eventually be responsible
for setting up the initial state of a level,
for enabling and disabling appropriate objects
when the Switch button is pressed, and positioning
the player at the start location.

PlayerMain now has a delegate,
DimensionButtonPressed, that
LevelManager.SwitchDimensions is subscribed to. It
is invoked when the E key is pressed.