Modify LevelManager and DimensionSwitch
DimensionSwitch has a new method, ChangeLayout,
that will eventually enable and disable the
relevant elements. At the moment, it only
prints a debug message.

LevelManager.SwitchDimensions has a integer
variable, _currentLevel, that holds the current
level number. There is an array, LevelObjects,
that contains the DimensionSwitch objects of
all of the levels.
SwitchDimensions will call ChangeLayout for
the object in  LevelObjects[_currentLevel]