Modify DimensionSwitch
DimensionSwitch has a new method, ToggleTileMaps.
This method first enables the tilemap that's being
"switched" to, then disables the tilemap that's
being "switched" from, and finally sets the index
pointer value for the next switch.

This is done to prevent the player from sinking
into the terrain should the enabled terrain be
disabled first.