Modify EndZone, LevelManager, and DimensionSwitch
EndZone has a new delegate, LevelCompleted. It
requires an integer as an input. This is invoked
after EndZoneEntered is invoked.

Endzone also has a new integer variable,
NextLevel, and a boolean variable, IsFinal.
NextLevel determines which level should be
played next. IsFinal will determine whether
that level is the final level. There will be a
method that will cause the delegate to be invoked
such that the player is returned to the title menu.

LevelManager now has a new method, SetCurrentLevel.
This is subscribed to EndZone.LevelCompleted. This
will set _currentLevel to the next target level.