Modify PlayerMain, LevelManager, DimensionSwitch, and UIManager
UIManager.PlayerGame is now renamed to
ContinueNextLevel. This also takes on the role of
invoking NextLevelRequested.

LevelManager.StartNextLevel now disables the previous
level's elements, enables the next level's elements,
and calls the setup method for the next level.

DimensionSwitch.SetupLevel now sets up the default
starting elements for the level. This also includes
setting the player's position by invoking the new
delegate, LevelStarted. LevelStarted requires a
Vector3 value.

PlayerMain has a new method, RepositionPlayer. This
sets the player's position to whatever value is passed
through. This method is subscribed to
DimensionSwitch.SetupLevel.

There's an issue with the player not jumping at all.