Fix for player not jumping
Issue was caused by dimensions of sprite.
The height of the sprite was being used to calculate the
raycast's origin in IsGrounded method. However, the
clear areas of the sprite were not sliced out and were
also included as part of the sprite's height. This
was causing the raycast origin to be further underneath
the player sprite. Current fix was to edit the sprite
and slice off the clear sections surrounding the sprite.