Modify LevelManager and UIManager
UIManager has a new delegate, IntValueRequested.
This delegate returns an int value.

ContinueToNextLevel has been modified so that
on level 0, the title canvas is enabled. This
invokes IntValueRequested for the level number.

TriggerTitleCanvas has been modified so that the
player can navigate through the title screen level.
The intention is to make the title menu a playable
menu.

LevelManager has new method, GetCurrentLevel.
This returns the current level number. This is
subscribed to UIManager.IntValueRequested.