Modify LevelManager
Added Camera variable, and a new method, SetCameraSize.
SetCameraSize changes the orthographic size of the camera.
The level sizes may vary, so the camera size should change
to reduce the amount of empty space being shown.

An array, CameraOrthSizes, will hold the sizes of the
camera for each level. It will set the size to 20
if there isn't a corresponding index for the current
level number.