Add Water SFX and WaterCurrent class
Imported Water Spash Pack from unity Asset store.
Added a sound effect for entering water and thorns.

Added WaterCurrent class. Originally meant for only
the water flow, it will now be used for any obstacle
that is meant to apply a force to the player.
It will apply a force to the player rigid body in a
specific direction. The direction is set from the
inspector. It will also play a sound effect based
on the gameobject's tag name.

AudioManager is now subscribed to
WaterCurrent.PlaySFX.