Modify level one and PlayerMain
Add WaterCurrent script to thorns. Modified Tilemap
layout so that thorns upward force will cause player
to be "caught" in the ceiling, necessitating the
player to "switch" to the oasis to remove the
thorns.

PlayerMain jump check modified: _isAirborne check
removed. It is suspected that this was causing the
issue where you sometimes were not able to jump
after jumping and having a side collision.